+++ src/main/kotlin/com/protokotlin/gradle/ProtoKotlinExtension.kt
@@ -1,5 +1,6 @@
 package com.protokotlin.gradle
 
+import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.provider.Property
 
@@ -10,4 +11,10 @@ abstract class ProtoKotlinExtension {
     abstract val outputDir: DirectoryProperty
     
     abstract val packageName: Property<String>
+    
+    /**
+     * Additional directories to search for imported proto files.
+     * Similar to protoc's -I/--proto_path option.
+     */
+    abstract val protoPath: ConfigurableFileCollection
 }